Remove old `find_python` functionality
Move python version stuff into new_python_linter bc it is not used anywhere
else.
